Refactor the API to be more scalable
