--- modules/api/src/endpoints/GET/index.js
@@ -1,8 +1,0 @@
-export { default as address    } from './address'
-export { default as login      } from './login'
-export { default as apikeys    } from './apikeys'
-export { default as merchantid } from './merchantid'
-export { default as username   } from './username'
-export { default as payments   } from './payments'
-export { default as totalsales } from './totalsales'
-export { default as xpub       } from './xpub'
