--- modules/api/src/endpoints/POST/index.js
@@ -1,6 +1,0 @@
-export { default as register } from './register'
-export { default as address  } from './address'
-export { default as pay     } from './pay'
-export { default as paid    } from './paid'
-export { default as password } from './password'
-export { default as username } from './username'
