+++ modules/api/src/endpoints/api/index.js
@@ -1,0 +1,1 @@
+export { default as keys } from './keys'
