+++ modules/api/src/endpoints/index.js
@@ -1,0 +1,2 @@
+export { default as pay  } from './pay'
+export { default as paid } from './paid'
