+++ modules/api/src/endpoints/user/password.js
