+++ modules/api/src/endpoints/user/payout/index.js
@@ -1,0 +1,3 @@
+export { default as address } from './address'
+export { default as xpub    } from './xpub'
+export { default as method  } from './method'
