+++ modules/api/src/endpoints/user/register.js
