+++ modules/api/src/endpoints/user/sales.js
@@ -1,13 +1,13 @@
 /**
- * GET /username API endpoint
+ * GET /totalsales API endpoint
  * @author The Gateway Project Developers <hello@gateway.cash>
- * @file Defines a GET endpoint for /username
+ * @file Defines a GET endpoint for /totalsales
  */
 import { mysql, handleResponse, auth } from 'utils'
 import url from 'url'
 
-export default async (req, res) => {
-  console.log('GET /username requested')
+let GET = async (req, res) => {
+  console.log('GET /totalsales requested')
 
   // parse the provided data
   const query = url.parse(req.url, true).query
@@ -17,10 +17,14 @@ export default async (req, res) => {
   if (!userIndex) return
 
   let result = await mysql.query(
-    'SELECT username FROM users WHERE tableIndex = ? LIMIT 1',
+    'SELECT totalSales FROM users WHERE tableIndex = ? LIMIT 1',
     [userIndex]
   )
   return handleResponse({
-    username: result[0].username
+    totalSales: result[0].totalSales
   }, res)
+}
+
+export default {
+  GET: GET
 }
