+++ modules/api/src/main.js
@@ -6,28 +6,14 @@
 import express from 'express'
 import bodyParser from 'body-parser'
 import dotenv from 'dotenv'
+import { fundsTransferService, brokenPaymentsService } from 'services'
+import tests from 'tests'
+import routes from 'routes'
 dotenv.config()
 
-import * as GETEndpoints from 'endpoints/GET'
-import * as POSTEndpoints from 'endpoints/POST'
-
-// include all service daemons
-import { fundsTransferService, brokenPaymentsService } from 'services'
-
-// include test scripts
-import tests from 'tests'
-
-// print startup message
-console.log('Starting Web Services Backend...')
-
-// create the express instance
 const app = express()
-
-// se up bodyParser tor form data handling
 app.use(bodyParser.json())
 app.use(bodyParser.urlencoded({ extended: true }))
-app.use(express.json())
-app.use(express.urlencoded({ extended: true }))
 
 // enable CORS to allow for API calls from other sites
 app.use((req, res, next) => {
@@ -39,6 +25,7 @@ app.use((req, res, next) => {
   next()
 })
 
+// consolidate form and request data into one place
 app.use((req, res, next) => {
   req.body = {...req.body, ...req.params, ...req.query}
   next()
@@ -49,12 +36,7 @@ app.use((req, res, next) => {
 app.use(express.static('public'))
 
 // utilize the API endpoints for appropriate requests
-Object.keys(GETEndpoints).forEach((e) => {
-  app.get('/' + e, GETEndpoints[e])
-})
-Object.keys(POSTEndpoints).forEach((e) => {
-  app.post('/' + e, POSTEndpoints[e])
-})
+routes(app)
 
 // listen for connections
 app.listen(process.env.WEB_PORT, () => {
@@ -66,7 +48,7 @@ if (!process.env.TEST_MODE) {
   // run the main processor every 30 seconds
   setInterval(fundsTransferService, 30000)
   // run the broken payments processor every 12 hours
-  setInterval(brokenPaymentsService.run, 43200000)
+  setInterval(brokenPaymentsService, 43200000)
 
 // run tests if we are in test mode
 } else {
